requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-Login==0.6.3
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.0.2
opencv-python==4.12.0.88
pillow==11.3.0
pymongo==4.15.1
Werkzeug==3.1.3
zipp==3.21.0
